requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
mlflow>=2.0.0
pyyaml>=6.0
joblib>=1.1.0
